% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_art_gift_load.R
\name{map_art_gift_load}
\alias{map_art_gift_load}
\title{Load the data}
\usage{
map_art_gift_load(
  region,
  name_place,
  crs = 3785,
  dist_shape = 3,
  shape = "circle"
)
}
\arguments{
\item{region}{The name of the directory where you can find the shp}

\item{name_place}{The name of the place to plot}

\item{crs}{The EPSG CRS, default to Mercator (3785)}

\item{dist_shape}{The distance to plot around the plot}

\item{shape}{The shape, circle or square, default to circle}
}
\value{
The loaded shp in the GlobalEnv
}
\description{
Load the data
}
