% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_art_gift_plot.R
\name{map_art_gift_plot}
\alias{map_art_gift_plot}
\title{Plot the map}
\usage{
map_art_gift_plot(
  region,
  name_place,
  size_caption = 120,
  width_out = 297,
  height_out = 420,
  place_shape = place_shape,
  roads = roads_cleaned,
  water = water_cropped,
  railways = railways_cropped,
  forest = forest,
  building = building_cropped
)
}
\arguments{
\item{region}{The name of the directory where you can find the shp}

\item{name_place}{The name of the place to plot}

\item{size_caption}{The size of the caption}

\item{width_out}{Output width in mm}

\item{height_out}{Output height in mm}

\item{place_shape}{The shp of the place, default in GlobalEnv}

\item{roads}{The shp of the roads, default in GlobalEnv}

\item{water}{The shp of the water, default in GlobalEnv}

\item{railways}{The shp of the railways, default in GlobalEnv}

\item{forest}{The shp of the forest, default in GlobalEnv}

\item{building}{The shp of the building, default in GlobalEnv}
}
\value{
A saved plot in maps/output
}
\description{
Plot the map
}
